Create Rule enum
Primary purpose of this commit is to introduce a `Rule` enum to the
quadrature integration routine.
This allows the replacement trait `Integrate`, the implementing types
`GaussKronrodX`, and the types `SharedData` and `ExtendedData` to be
private implementation details.
The integration rule used in the `Basic` and `Adaptive` routines is now
imposed via selecting a variant of the enum `Rule`.

- Creates `Rule` enum
- Renames `Rule` trait to `Integrate`
- Changes visibility of `Integrate`, `GaussKronrodX`, `SharedData`,
  `ExtendedData`
- Updates unit and integration tests.
- Introduces constructor methods on `IntegralEstimate` so that estimates
  can be obtained via, e.g. `IntegralEstimate::adaptive(...)`