Change implementation of integration rules
The integration rules for Gauss-Kronrod quadrature integration were
previously implemented in a cumbersome way via a trait `Integrator`
implemented on unit structs `GaussKronrodN` for an N-point integration
rule, with a `Rule` enum used to select the appropriate rule by the
user. This has now been significantly simplified to the use of a struct
`Rule` which contains as fields references to the constant arrays of
evaluation points and weights necessary for the integration routine.

CHANGES:

- Remove `Integrator` trait.
- Remove `Rule` enum.
- Remove `GaussKronrodN` unit structs.
- Add `Rule` struct and generators `Rule::gkN()` for N-point
  Gauss-Kronrod integration rules.
- Add `GaussKronrodIntegrator` stuct for internal integration.
- Standardise API's between `Basic`, `Adaptive` and
  `AdaptiveSingularity` integrators.
- Change field ordering in constructors.
- Fix tests to incorporate above changes.