Add complex function eval. to basic integrator
This initial commit adds the capability of integrating complex valued
functions of a real variable to the simple Integrator routine.

This is achieved by adding an associated type to the Integrand trait:

trait Integrand {
    type Scalar: ScalarF64;
    fn evaluate(&self, x: f64) -> Self::Scalar;
}

where ScalarF64 is a sealed trait implemented for std::f64 and
num_complex::Complex<f64>. The function to be integrated is of a real
variable, x: f64, but returns complex values when evaluated.