[WIP] Add multi::Adaptive integration routine
This commit adds the basic functionality of the adaptive integration
routine for multi-dimensional integrals, exposing the multi::Adaptive
type.

Includes a simple test of the routine which calculates the integral
result to a relative difference of < 1e-15 compared to the equivalent
dchure result, BUT the approximated error has only a relative difference of
< 1e-1 compared to the dcuhre approximated error. Termination of the
integration routine occurs in the same number of iterations in both
multi::Adaptive and dcuhre.

TODO: Check error estimation for bugs.