Move error estimate adjustment to Region::bisect
Previous implementation of the error adjustment was done inside the
multi::Workspace::improved_result_error method. This required regions
calculated in the adaptive routine to be mutable and separated the
creation of the "correct" region estimate from the actual integration
step. This commit moves the error adjustment to the Region::bisect
method, which keeps the mutability contained only to this function.